Deployed to Test: #5196

Related work items: #5196

+++ ERP.Web.API/Domain/Services/Accounting/JournalService.cs
@@ -22,7 +22,7 @@ namespace ERP.Web.API.Domain.Services.Accounting
             _db = db;
         }
 
-        public SaveResult PostingJournal(JournalRequest data)
+        public SaveResult PostingJournal(JournalRequest data, int userId)
         {
             var result = new SaveResult(false);
             var systemParam = _db.SystemParameters.ToList();
@@ -32,6 +32,12 @@ namespace ERP.Web.API.Domain.Services.Accounting
             using var transaction = _db.Database.BeginTransaction();
             try
             {
+                if (_db.PostingLogs.Where(x => Convert.ToInt32(x.Period) < Convert.ToInt32(data.Date.ToString("yyyyMM")) && x.IsPosted == false).Any())
+                {
+                    result.Message = "Tidak bisa melakukan posting jurnal karena terdapat periode sebelumnya yang belum diposting.";
+                    return result;
+                }
+
                 var removed = _db.Journals.Where(x => x.Date.Month == data.Date.Month && x.Date.Year == data.Date.Year).ToList();
                 if (removed != null)
                     _db.RemoveRange(removed);
@@ -110,6 +116,13 @@ namespace ERP.Web.API.Domain.Services.Accounting
                     if (journalEY != null)
                         _db.AddRange(journalEY);
                 }
+
+                //Update posting log
+                var plData = _db.PostingLogs.FirstOrDefault(x => x.Period == data.Date.ToString("yyyyMM"));
+                plData.IsPosted = true;
+                plData.PostedBy = userId;
+                plData.PostedDate = DateTime.Now;
+                _db.PostingLogs.Update(plData);
 
                 _db.SaveChanges();
 
@@ -325,6 +338,8 @@ namespace ERP.Web.API.Domain.Services.Accounting
                 {
                     var smData = _db.StockMutations.FirstOrDefault(x => x.RefDetailId1 == itemDetail.DlvDetail.Id && x.RefCode1 == itemDetail.DlvDetail.Code);
                     var prorateHeaderDisc = itemData.Dlvheader.FinalDisc > 0 ? (itemData.Dlvheader.FinalDisc * itemDetail.DlvDetail.NettPrice) / DlvDetailData.Sum(x => x.DlvDetail.NettPrice) : 0;
+                    var nonVoidSM = RemoveVoidSM(_db.StockMutations.ToList());
+                    var resultHpp = CalculateHPP(nonVoidSM, smData.WarehouseCode, smData.ItemId, smData.RefDetailId1);
 
                     //Discount - Diskon
                     if (itemDetail.DlvDetail.Disc > 0)
@@ -388,7 +403,7 @@ namespace ERP.Web.API.Domain.Services.Accounting
                         CurrCode = itemData.Dlvheader.CurrCode,
                         Period = itemData.Dlvheader.Date.ToString("yyyyMMdd"),
                         Type = "C",
-                        Amount = ivnValue,
+                        Amount = resultHpp > 0 ? resultHpp * smData.BaseQty : 0,
                         SrcTrans = "DLV"
                     });
 
@@ -407,7 +422,7 @@ namespace ERP.Web.API.Domain.Services.Accounting
                         CurrCode = itemData.Dlvheader.CurrCode,
                         Period = itemData.Dlvheader.Date.ToString("yyyyMMdd"),
                         Type = "D",
-                        Amount = smData != null ? smData.BaseNettPrice * smData.BaseQty : 0,
+                        Amount = resultHpp > 0 ? resultHpp * smData.BaseQty : 0,
                         SrcTrans = "DLV"
                     });
 
@@ -1977,6 +1992,10 @@ namespace ERP.Web.API.Domain.Services.Accounting
                 short l = 0;
                 foreach (var itemDetail in adjDetailData)
                 {
+                    var smData = _db.StockMutations.FirstOrDefault(x => x.RefDetailId1 == itemDetail.AdjDetail.Id && x.RefCode1 == itemDetail.AdjDetail.Code);
+                    var nonVoidSM = RemoveVoidSM(_db.StockMutations.ToList());
+                    var resultHpp = CalculateHPP(nonVoidSM, smData.WarehouseCode, smData.ItemId, smData.RefDetailId1);
+
                     if (itemDetail.AdjDetail.QtyAdjust > itemDetail.AdjDetail.QtyOnHand)
                     {
                         var qty = (itemDetail.AdjDetail.QtyAdjust - itemDetail.AdjDetail.QtyOnHand);
@@ -1993,7 +2012,7 @@ namespace ERP.Web.API.Domain.Services.Accounting
                             CurrCode = "IDR",
                             Period = itemData.Date.ToString("yyyyMMdd"),
                             Type = "D",
-                            Amount = qty * itemDetail.AdjDetail.COGS,
+                            Amount = qty * resultHpp,
                             SrcTrans = "ADJ"
                         });
                     }
@@ -2013,7 +2032,7 @@ namespace ERP.Web.API.Domain.Services.Accounting
                             CurrCode = "IDR",
                             Period = itemData.Date.ToString("yyyyMMdd"),
                             Type = "C",
-                            Amount = qty * itemDetail.AdjDetail.COGS,
+                            Amount = qty * resultHpp,
                             SrcTrans = "ADJ"
                         });
                     }
@@ -2098,6 +2117,75 @@ namespace ERP.Web.API.Domain.Services.Accounting
             }
 
             return journals;
+        }
+
+        private decimal CalculateHPP(IEnumerable<StockMutation> stockMutations, string whCode, int itemId, long id)
+        {
+            decimal latestQty = 0;
+            decimal latestStockValue = 0;
+            decimal hpp = 0;
+
+            var firstId = stockMutations.FirstOrDefault(x => x.WarehouseCode == whCode && x.ItemId == itemId && x.Src == "RCV")?.Id;
+            if (firstId == null)
+                return 0;
+            var firstSM = stockMutations.FirstOrDefault(x => x.Id == firstId);
+
+            var currentSM = stockMutations.FirstOrDefault(x => x.WarehouseCode == whCode && x.ItemId == itemId && x.RefDetailId1 == id);
+
+            if (firstSM != null || firstSM.BaseNettPrice != 0)
+            {
+                latestStockValue += firstSM.BaseNettPrice * firstSM.BaseQty;
+                latestQty += firstSM.BaseQty;
+                hpp = latestStockValue / latestQty;
+
+                var listSM = stockMutations.Where(x => new[] { "RCV", "DO", "SR", "ADJ", "TS" }.Contains(x.Src) && x.WarehouseCode == whCode && x.ItemId == itemId && x.Id > firstId && x.Id <= currentSM.Id).OrderBy(x => x.Date).ToList();
+                foreach (var item in listSM)
+                {
+                    if (item.Src == "RCV" && item.Src == "SR")
+                    {
+                        latestStockValue += item.BaseNettPrice * item.BaseQty;
+                        latestQty += item.BaseQty;
+                    }
+                    else //if (item.Src == "DO")
+                    {
+                        var srcTrans = stockMutations.FirstOrDefault(x => x.ItemId == item.ItemId && x.RefCode1 == item.RefCode2);
+                        if (srcTrans?.Src == "SR")
+                        {
+                            item.BaseNettPrice = srcTrans.BaseNettPrice;
+                            item.NettPrice = srcTrans.NettPrice;
+                            latestStockValue -= item.BaseNettPrice * item.BaseQty;
+                            latestQty -= item.BaseQty;
+                        }
+                        else
+                        {
+                            item.BaseNettPrice = hpp;
+                            item.NettPrice = (hpp * item.BaseQty) / item.Qty;
+                            latestStockValue -= item.BaseNettPrice * item.BaseQty;
+                            latestQty -= item.BaseQty;
+                        }
+                        _db.StockMutations.Update(item);
+                    }
+                    if (latestStockValue > 0 && latestQty > 0)
+                        hpp = latestStockValue / latestQty;
+                }
+                _db.SaveChanges();
+            }
+            return hpp;
+        }
+
+        private List<StockMutation> RemoveVoidSM(List<StockMutation> data)
+        {
+            var result = data;
+            var listVoid = new List<string>();
+
+            listVoid.AddRange(_db.PurchaseReceiveHeaders.Where(x => x.Mark == "V").Select(x => x.Code).ToList());
+            listVoid.AddRange(_db.SalesDeliveryHeaders.Where(x => x.Mark == "V").Select(x => x.Code).ToList());
+            listVoid.AddRange(_db.AdjustmentHeaders.Where(x => x.Mark == "V").Select(x => x.Code).ToList());
+            listVoid.AddRange(_db.TransferStockHeaders.Where(x => x.Mark == "V").Select(x => x.Code).ToList());
+
+            result = result.Where(x => !listVoid.Contains(x.RefCode1)).ToList();
+
+            return result;
         }
     }
 }
